allow grpc without core
