+++ applications/minotari_app_grpc/Cargo.toml
@@ -7,10 +7,15 @@ authors.workspace = true
 repository.workspace = true
 license.workspace = true
 
+[features]
+default = ["base_node"]
+base_node = ["tari_core"]
+
+
 [dependencies]
 tari_common_types = { workspace = true }
 tari_comms = { workspace = true }
-tari_core = { workspace = true}
+tari_core = { workspace = true, optional = true}
 tari_crypto = { workspace = true }
 tari_script = { workspace = true }
 tari_node_components = { workspace = true }
@@ -22,10 +27,8 @@ tari_transaction_components = { workspace = true }
 argon2 = { version = "0.4.1", features = ["std", "password-hash"] }
 base64 = "0.13.0"
 borsh = "1.5.7"
-chrono = { version = "0.4", default-features = false }
 log = "0.4"
 prost = "0.13.3"
-prost-types = "0.13.3"
 rand = "0.8"
 rcgen = "0.12.1"
 subtle = "2.5.0"
