+++ applications/minotari_app_grpc/src/conversions/mod.rs
@@ -21,6 +21,7 @@
 // USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 pub mod aggregate_body;
+#[cfg(feature = "base_node")]
 pub mod base_node_state;
 pub mod block;
 pub mod block_header;
@@ -28,12 +29,15 @@ pub mod chain_metadata;
 pub mod com_and_pub_signature;
 pub mod commitment_signature;
 pub mod consensus_constants;
+#[cfg(feature = "base_node")]
 pub mod historical_block;
 pub mod new_block_template;
 pub mod output_features;
 pub mod peer;
 pub mod proof_of_work;
 pub mod sidechain_feature;
+#[cfg(feature = "base_node")]
+pub mod sidechain_validator;
 pub mod signature;
 pub mod transaction;
 pub mod transaction_input;
@@ -41,17 +45,7 @@ pub mod transaction_kernel;
 pub mod transaction_output;
 pub mod unblinded_output;
 
-use prost_types::Timestamp;
-
 use crate::{tari_rpc as grpc, tari_rpc::BlockGroupRequest};
-
-/// Utility function that converts a `chrono::NaiveDateTime` to a `prost::Timestamp`
-pub fn naive_datetime_to_timestamp(datetime: chrono::NaiveDateTime) -> Timestamp {
-    Timestamp {
-        seconds: datetime.and_utc().timestamp(),
-        nanos: 0,
-    }
-}
 
 impl From<u64> for grpc::IntegerValue {
     fn from(value: u64) -> Self {
